requirements.txt
discord.py
sqlalchemy
psycopg2-binary
python-dotenv
aiohttp
